+++ src/main/java/ru/yandex/practicum/filmorate/storage/film/FilmDbStorage.java
@@ -5,10 +5,7 @@ import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
-import ru.yandex.practicum.filmorate.exception.NotExistException;
 import ru.yandex.practicum.filmorate.model.Film;
-import ru.yandex.practicum.filmorate.service.DirectorService;
-import ru.yandex.practicum.filmorate.service.GenreService;
 
 import java.sql.Date;
 import java.sql.PreparedStatement;
@@ -16,19 +13,14 @@ import java.sql.Statement;
 import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
-import java.util.stream.Collectors;
 
 @Repository
 public class FilmDbStorage implements FilmStorage {
 
     private final JdbcTemplate jdbcTemplate;
-    private final GenreService genreService;
-    private final DirectorService directorService;
 
-    public FilmDbStorage(JdbcTemplate jdbcTemplate, GenreService genreService, DirectorService directorService) {
+    public FilmDbStorage(JdbcTemplate jdbcTemplate) {
         this.jdbcTemplate = jdbcTemplate;
-        this.genreService = genreService;
-        this.directorService = directorService;
     }
 
     @Override
@@ -50,38 +42,23 @@ public class FilmDbStorage implements FilmStorage {
         }, keyHolder);
 
         film.setId((Objects.requireNonNull(keyHolder.getKey()).longValue()));
-        genreService.addAll(film.getId(), film.getGenres());
-        directorService.addAllToFilm(film.getId(), film.getDirectors());
         return film;
     }
 
     @Override
     @Transactional
     public Film update(Film film) {
-        if (film == null) {
-            throw new NotExistException("Передан пустой аргумент!");
-        }
-        String sqlQuery = "UPDATE films SET " +
-                "name = ?, description = ?, release_date = ?, duration = ?, " +
-                "mpa_rating_id = ? WHERE id = ?";
-        if (jdbcTemplate.update(sqlQuery,
+        jdbcTemplate.update("UPDATE films " +
+                        "SET name = ?, description = ?, release_date = ?, duration = ?, mpa_rating_id = ? " +
+                        "WHERE id = ?",
                 film.getName(),
                 film.getDescription(),
                 film.getReleaseDate(),
                 film.getDuration(),
                 film.getMpa().getId(),
-                film.getId()) != 0) {
-            var genres = Optional.ofNullable(film.getGenres()).stream()
-                    .flatMap(List::stream)
-                    .distinct()
-                    .collect(Collectors.toList());
-            film.setGenres(genres);
-            genreService.update(film.getId(), genres);
-            directorService.updateAllToFilm(film.getId(), film.getDirectors());
-            return film;
-        } else {
-            throw new NotExistException("Фильм с ID=" + film.getId() + " не найден!");
-        }
+                film.getId());
+        return film;
+
     }
 
     @Override
@@ -91,7 +68,7 @@ public class FilmDbStorage implements FilmStorage {
     }
 
     @Override
-    public Film get(long id) {
+    public Optional<Film> get(long id) {
         return jdbcTemplate.query(
                         "SELECT f.id, f.name, f.description, f.release_date, f.duration, "
                                 + "f.mpa_rating_id, m.name AS mpa_name, COUNT(l.user_id) AS likes, "
@@ -110,8 +87,7 @@ public class FilmDbStorage implements FilmStorage {
                                 + "GROUP BY f.id "
                                 + "ORDER BY COUNT(l.user_id)",
                         new Object[]{id}, new FilmMapper()).stream()
-                .findFirst()
-                .orElse(null);
+                .findFirst();
     }
 
     @Override
