+++ src/main/java/ru/yandex/practicum/filmorate/storage/film/FilmStorage.java
@@ -3,6 +3,7 @@ package ru.yandex.practicum.filmorate.storage.film;
 import ru.yandex.practicum.filmorate.model.Film;
 
 import java.util.List;
+import java.util.Optional;
 
 public interface FilmStorage {
 
@@ -12,7 +13,7 @@ public interface FilmStorage {
 
     void delete(long id);
 
-    Film get(long id);
+    Optional<Film> get(long id);
 
     List<Film> findAll();
